# openapi-server/swagger_server/models/response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Response(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, message: str=None, status: str=None):  # noqa: E501
        """Response - a model defined in Swagger

        :param message: The message of this Response.  # noqa: E501
        :type message: str
        :param status: The status of this Response.  # noqa: E501
        :type status: str
        """
        self.swagger_types = {
            'message': str,
            'status': str
        }

        self.attribute_map = {
            'message': 'message',
            'status': 'status'
        }
        self._message = message
        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'Response':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Response of this Response.  # noqa: E501
        :rtype: Response
        """
        return util.deserialize_model(dikt, cls)

    @property
    def message(self) -> str:
        """Gets the message of this Response.


        :return: The message of this Response.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this Response.


        :param message: The message of this Response.
        :type message: str
        """

        self._message = message

    @property
    def status(self) -> str:
        """Gets the status of this Response.


        :return: The status of this Response.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this Response.


        :param status: The status of this Response.
        :type status: str
        """

        self._status = status