# openapi-server/swagger_server/swagger/swagger.yaml
openapi: 3.0.2
info:
  title: Authentication service
  version: "2.0"
servers:
- url: /
paths:
  /api/v1/assign-roles:
    post:
      tags:
      - ASSIGN_ROLES
      summary: Endpoint to assign roles to user
      description: Assign roles to user
      operationId: api_v1_assign_roles_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleRequest'
            example:
              role_ids:
              - a9c6e8da-f2bf-458a-978b-d2f50a031451
              - 7cf56926-054c-4522-ac6f-d9f5d0e9d18e
              user_id: 7cd483e9-5888-40fd-813a-a382154bcfd2
      responses:
        "201":
          description: Roles were assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: roles were assigned to user
                status: success
                user_roles:
                - id: 4a73b964-af72-4801-aed9-113783561540
                  role_id: a9c6e8da-f2bf-458a-978b-d2f50a031451
                  user_id: 7cd483e9-5888-40fd-813a-a382154bcfd2
                - id: 0f55b9d8-f027-4766-9476-2b89e17c1854
                  role_id: 7cf56926-054c-4522-ac6f-d9f5d0e9d18e
                  user_id: 7cd483e9-5888-40fd-813a-a382154bcfd2
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                norole:
                  value:
                    message: role not found
                    status: error
                nouser:
                  value:
                    message: user not found
                    status: error
      security:
      - jwt_auth:
        - write:admin
        - read:admin
      x-openapi-router-controller: swagger_server.controllers.assignroles_controller
...
  /api/v1/check-permissions:
    post:
      tags:
      - CHECK_PERMISSIONS
      summary: Endpoint to check user permissions
      description: check if user belongs to specified roles
      operationId: api_v1_check_permissions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleRequest'
            example:
              role_ids:
              - a9c6e8da-f2bf-458a-978b-d2f50a031451
              - 7cf56926-054c-4522-ac6f-d9f5d0e9d18e
              user_id: 7cd483e9-5888-40fd-813a-a382154bcfd2
      responses:
        "200":
          description: Roles were checked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
              examples:
                approved:
                  value:
                    has_permissions: true
                    message: roles were checked successfully
                    status: success
                disapproved:
                  value:
                    has_permissions: false
                    message: roles were checked successfully
                    status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                norole:
                  value:
                    message: role not found
                    status: error
                nouser:
                  value:
                    message: user not found
                    status: error
      security:
      - jwt_auth:
        - write:admin
        - read:admin
      x-openapi-router-controller: swagger_server.controllers.checkpermissions_controller
  /api/v1/role:
    get:
      tags:
      - ROLE_LIST
      summary: Endpoint to get all roles
      description: List of all available roles
      operationId: api_v1_role_get
      responses:
        "200":
          description: List of roles is available
          content:
            application/json:
              schema:
                type: array
                example:
                - code: admin
                  description: unlimited access to all actions
                  id: a9c6e8da-f2bf-458a-978b-d2f50a031451
                - code: subscriber
                  description: account without paying for registered users
                  id: 7cf56926-054c-4522-ac6f-d9f5d0e9d18e
                - code: member
                  description: account with payment options
                  id: 7166fd5f-a4e4-45f0-952c-78d0297c7b03
                items:
                  $ref: '#/components/schemas/Role'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      x-openapi-router-controller: swagger_server.controllers.rolelist_controller
    post:
      tags:
      - CREATE_ROLE
      summary: Endpoint to create new role
      description: Create new role
      operationId: api_v1_role_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
            example:
              code: admin
              description: unlimited access to all actions
      responses:
        "201":
          description: List of roles is available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: New role was created
                role:
                  code: admin
                  description: unlimited access to all actions
                  id: a9c6e8da-f2bf-458a-978b-d2f50a031451
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
      security:
      - jwt_auth:
        - write:admin
        - read:admin
      x-openapi-router-controller: swagger_server.controllers.createrole_controller
  /api/v1/role/{role_id}:
    get:
      tags:
      - ROLE_DETAILS
      summary: Get role detailes
      description: detailed info about role
      operationId: api_v1_role_role_id_get
      parameters:
      - name: role_id
        in: path
        description: Role uuid
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example:
          role_id: a9c6e8da-f2bf-458a-978b-d2f50a031451
      responses:
        "200":
          description: info about role is available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: info about role is available
                role:
                  code: admin
                  description: unlimited access to all actions
                  id: a9c6e8da-f2bf-458a-978b-d2f50a031451
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      x-openapi-router-controller: swagger_server.controllers.roledetails_controller
    delete:
      tags:
      - DELETE_ROLE
      summary: Endpoint to delete role
      description: delete role
      operationId: api_v1_role_role_id_delete
      parameters:
      - name: role_id
        in: path
        description: Role uuid
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example:
          role_id: a9c6e8da-f2bf-458a-978b-d2f50a031451
      responses:
        "204":
          description: role was deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: role was deleted successfully
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - write:admin
        - read:admin
      x-openapi-router-controller: swagger_server.controllers.deleterole_controller
    patch:
      tags:
      - CHANGE_ROLE_DETAILS
      summary: Endpoint to change role
      description: change role info
      operationId: api_v1_role_role_id_patch
      parameters:
      - name: role_id
        in: path
        description: Role uuid
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example:
          role_id: a9c6e8da-f2bf-458a-978b-d2f50a031451
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
            example:
              code: admin
              description: unlimited access to all actions
      responses:
        "200":
          description: info about role was changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: info about role was changed successfully
                role:
                  code: admin
                  description: unlimited access to all actions
                  id: a9c6e8da-f2bf-458a-978b-d2f50a031451
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      x-openapi-router-controller: swagger_server.controllers.changeroledetails_controller
components:
  schemas:
...
    Response:
      title: Response
      type: object
      properties:
        message:
          title: response message
          type: string
        status:
          title: response status
          type: string
      example:
        message: message
        status: status
    Role:
      title: Role
      required:
      - code
      - description
      properties:
        code:
          title: role name
          type: string
        description:
          title: role detailed description
          type: string
        id:
          title: role uuid
          type: string
...
    UserRoleRequest:
      title: UserRoleRequest
      properties:
        role_ids:
          type: array
          items:
            title: role uuids
            type: string
        user_id:
          title: user uuid
          type: string
...
    inline_response_200:
      properties:
        has_permissions:
          type: boolean
        message:
          type: string
        status:
          type: string
      example:
        has_permissions: true
        message: message
        status: status
...
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_jwt_auth